+++ app/components/time-ctx/time-ctx-directive.js
@@ -40,8 +40,6 @@ angular.module('time-ctx')
         + GRAPH_5_6th_PADDING_RATIO * UtilService.TIMELINE_LEFT_MARGIN;
     };
 
-    Timeline.onresize = resize;
-
     scope.tctx.dims = {
       width: UtilService.getCurrentWidth()
         + GRAPH_5_6th_PADDING_RATIO * UtilService.TIMELINE_LEFT_MARGIN,
@@ -73,7 +71,6 @@ angular.module('time-ctx')
       angular.forEach(sharedKeys, function (key) {
         item[key] = response[key];
       });
-
       scope.tctx.content[response.layerSlug] = item;
     };
 
@@ -109,10 +106,15 @@ angular.module('time-ctx')
         : State.spatial.here;
 
     var getTimeData = function () {
+      var graphWidth = scope.tctx.dims.width -
+        scope.tctx.dims.padding.left -
+        scope.tctx.dims.padding.right;
+
       DataService.getData('time', {
         geom: geom,
         start: State.temporal.start,
         end: State.temporal.end,
+        minPoints: graphWidth,
         temporalOnly: true // TODO: actually implement this in data-service.
       }).then(null, null, function (response) {
 
@@ -132,6 +134,7 @@ angular.module('time-ctx')
         } else if (response.layerSlug === 'timeseries') {
           angular.forEach(response.data, function (ts) {
             ts.layerSlug = ts.name;
+            ts.type = response.layerSlug;
             putDataOnScope(ts);
           });
         } else {
@@ -161,7 +164,10 @@ angular.module('time-ctx')
 
     var applyResize = function () {
       scope.$apply(resize(tlDims));
+      getTimeData();
     };
+
+    Timeline.onresize = resize;
 
     window.addEventListener('resize', applyResize);